[DEV-1954] Refactoring and add migration
