+++ src/subdomains/supporting/payment/repositories/blockchain-fee.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { BaseRepository } from 'src/shared/repositories/base.repository';
 import { EntityManager } from 'typeorm';
-import { BlockchainFee } from '../entities/blockchain-fee';
+import { BlockchainFee } from '../entities/blockchain-fee.entity';
 
 @Injectable()
 export class BlockchainFeeRepository extends BaseRepository<BlockchainFee> {
